requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
black==22.3.0
    # via -r requirements.in
click==8.1.3
    # via black
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
pyyaml==6.0
    # via -r requirements.in
smmap==5.0.0
    # via gitdb
tomli==2.0.1
    # via black
